+++ pg.el
@@ -1,9 +1,9 @@
 ;;; pg.el --- Socket-level interface to the PostgreSQL database  -*- lexical-binding: t -*-
 
-;; Copyright: (C) 1999-2002, 2022-2024  Eric Marsden
+;; Copyright: (C) 1999-2002, 2022-2025  Eric Marsden
 
 ;; Author: Eric Marsden <eric.marsden@risk-engineering.org>
-;; Version: 0.45
+;; Version: 0.46
 ;; Keywords: data comm database postgresql
 ;; URL: https://github.com/emarsden/pg-el
 ;; Package-Requires: ((emacs "28.1") (peg "1.0"))